requirements.txt
pyinquirer
clanimate
pycryptodome
colorama
tabulate
pyfiglet